requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.1.0

# Database
psycopg2-binary==2.9.9

# Authentication & Security
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0

# AI/ML Libraries
scikit-learn==1.3.2
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.4
joblib==1.3.2

# OCR
pytesseract==0.3.10
opencv-python==4.8.1.78

# File Processing
python-magic==0.4.27
PyPDF2==3.0.1

# Web Server
gunicorn==21.2.0
whitenoise==6.6.0

# Development
python-decouple==3.8
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Utilities
celery==5.3.4
redis==5.0.1
requests==2.31.0
